LayoutTests/fast/events/drag-and-drop-autoscroll-inner-frame-expected.txt

For manual testing, drag and drop "Drop Me" downwards and then upwards.
Check autoscroll within an inner frame by drag-and-drop

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS Autoscroll should have scrolled the iframe downwards, and did.
PASS iframe.contentDocument.body.scrollTop < middleTermScrollOffset is true
